串資料庫成功 顯示菜單照片

+++ BackEnd/src/main/java/com/example/bistro/backstage/comment/Comment.java
@@ -41,7 +41,7 @@ public class Comment {
 	private String  commentProduct;
 	
 
-	private Integer commentRating;
+	private Short commentRating;
 	
 	
 	private String  commentMessage;
@@ -104,12 +104,12 @@ public class Comment {
 	}
 
 
-	public Integer getCommentRating() {
+	public Short getCommentRating() {
 		return commentRating;
 	}
 
 
-	public void setCommentRating(Integer commentRating) {
+	public void setCommentRating(Short commentRating) {
 		this.commentRating = commentRating;
 	}
